+++ app/Widgets/ChatActions/ChatActions.php
@@ -8,7 +8,6 @@ use App\Widgets\Chat\Chat;
 use App\Widgets\ContactActions\ContactActions;
 use App\Widgets\Dialog\Dialog;
 use App\Widgets\Drawer\Drawer;
-use App\Widgets\Toast\Toast;
 use Moxl\Xec\Action\Blocking\Block;
 use Moxl\Xec\Action\Blocking\Unblock;
 use Moxl\Xec\Action\Message\Moderate;
@@ -30,13 +29,13 @@ class ChatActions extends \Movim\Widget\Base
 
     public function onBlock(Packet $packet)
     {
-        Toast::send($this->__('blocked.account_blocked'));
+        $this->toast($this->__('blocked.account_blocked'));
         $this->rpc('Chat_ajaxGet', $packet->content);
     }
 
     public function onUnblock(Packet $packet)
     {
-        Toast::send($this->__('blocked.account_unblocked'));
+        $this->toast($this->__('blocked.account_unblocked'));
         $this->rpc('Chat_ajaxGet', $packet->content);
     }
 
@@ -162,7 +161,7 @@ class ChatActions extends \Movim\Widget\Base
 
     public function ajaxCopiedMessageText()
     {
-        Toast::send($this->__('chatactions.copied_text'));
+        $this->toast($this->__('chatactions.copied_text'));
     }
 
     /**
@@ -241,17 +240,19 @@ class ChatActions extends \Movim\Widget\Base
 
         if ($message && $message->resolved == false) {
             try {
-                $url = new Url;
-                $url->resolve(htmlspecialchars_decode(trim($message->body)));
-                $message->urlid = $url->id;
+                $url = Url::resolve(htmlspecialchars_decode(trim($message->body)));
 
-                if ($url->file) {
-                    $messageFile = $url->file;
-                    $messageFile->message_mid = $message->mid;
-                    $messageFile->save();
+                if ($url) {
+                    $message->urlid = $url->id;
+
+                    if ($url->messageFile) {
+                        $messageFile = $url->messageFile;
+                        $messageFile->message_mid = $message->mid;
+                        $messageFile->save();
+                    }
+
+                    $this->rpc('Chat.refreshMessage', $message->mid);
                 }
-
-                $this->rpc('Chat.refreshMessage', $message->mid);
             } catch (\Exception $e) {
                 logError($e);
             }
@@ -267,10 +268,10 @@ class ChatActions extends \Movim\Widget\Base
     public function ajaxHttpResolveUrl(string $url)
     {
         try {
-            $embed = (new Url)->resolve(trim($url));
+            $url = Url::resolve(trim($url));
 
-            if ($embed != null) {
-                $this->rpc('MovimTpl.fill', '#embed', (new Chat)->prepareEmbed($embed));
+            if ($url != null) {
+                $this->rpc('MovimTpl.fill', '#embed', (new Chat)->prepareEmbed($url));
             }
         } catch (\Exception $e) {
         }