+++ app/Widgets/CommunityConfig/CommunityConfig.php
@@ -4,7 +4,6 @@ namespace App\Widgets\CommunityConfig;
 
 use App\Widgets\Dialog\Dialog;
 use App\Widgets\Drawer\Drawer;
-use App\Widgets\Toast\Toast;
 use Movim\Image;
 use Movim\Librairies\XMPPtoForm;
 use Movim\Widget\Base;
@@ -46,19 +45,19 @@ class CommunityConfig extends Base
     public function onAvatarSet(Packet $packet)
     {
         $this->rpc('Dialog_ajaxClear');
-        Toast::send($this->__('avatar.updated'));
+        $this->toast($this->__('avatar.updated'));
     }
 
     public function onConfigSaved()
     {
-        Toast::send($this->__('communityaffiliation.config_saved'));
+        $this->toast($this->__('communityaffiliation.config_saved'));
     }
 
     public function onConfigError(Packet $packet)
     {
-        Toast::send(
+        $this->toast(
             $packet->content ??
-            $this->__('communityaffiliation.config_error')
+                $this->__('communityaffiliation.config_error')
         );
     }
 
@@ -70,8 +69,8 @@ class CommunityConfig extends Base
 
         $view = $this->tpl();
         $view->assign('info', \App\Info::where('server', $origin)
-                                       ->where('node', $node)
-                                       ->first());
+            ->where('node', $node)
+            ->first());
 
         Dialog::fill($view->draw('_communityconfig_avatar'));
     }
@@ -82,10 +81,10 @@ class CommunityConfig extends Base
             return;
         }
 
-        $key = $origin.$node.'avatar';
+        $key = $origin . $node . 'avatar';
 
         $p = new Image;
-        $p->fromBase($form->photobin->value);
+        $p->fromBase64($form->photobin->value);
         $p->setKey($key);
         $p->save(false, false, 'jpeg', 60);
 
@@ -94,10 +93,10 @@ class CommunityConfig extends Base
 
         $r = new AvatarSet;
         $r->setTo($origin)
-          ->setNode($node)
-          ->setUrl(Image::getOrCreate($key, false, false, 'jpeg', true))
-          ->setData($p->toBase())
-          ->request();
+            ->setNode($node)
+            ->setUrl(Image::getOrCreate($key, false, false, 'jpeg', true))
+            ->setData($p->toBase())
+            ->request();
     }
 
     public function ajaxGetConfig($origin, $node, $advanced = false)
@@ -108,7 +107,7 @@ class CommunityConfig extends Base
 
         $r = new GetConfig;
         $r->setTo($origin)
-          ->setNode($node);
+            ->setNode($node);
 
         if ($advanced) {
             $r->enableAdvanced();
@@ -125,8 +124,8 @@ class CommunityConfig extends Base
 
         $r = new SetConfig;
         $r->setTo($origin)
-          ->setNode($node)
-          ->setData(formToArray($data))
-          ->request();
+            ->setNode($node)
+            ->setData(formToArray($data))
+            ->request();
     }
 }