+++ app/Widgets/CommunityHeader/CommunityHeader.php
@@ -3,7 +3,6 @@
 namespace App\Widgets\CommunityHeader;
 
 use App\Widgets\Dialog\Dialog;
-use App\Widgets\Toast\Toast;
 use Movim\Widget\Base;
 
 use Moxl\Xec\Action\Disco\Request;
@@ -46,13 +45,15 @@ class CommunityHeader extends Base
 
     public function tonTestPublish(Packet $packet)
     {
-        list($origin, $node) = array_values($packet->content);
-        $this->rpc('MovimUtils.redirect', $this->route('publish', [$origin, $node]));
+        $this->rpc('MovimUtils.redirect', $this->route(
+            'publish',
+            [$packet->content->to, $packet->content->node]
+        ));
     }
 
     public function tonTestPublishError(Packet $packet)
     {
-        Toast::send($this->__('publish.no_publication'));
+        $this->toast($this->__('publish.no_publication'));
     }
 
     public function onSubscribed(Packet $packet)
@@ -61,12 +62,12 @@ class CommunityHeader extends Base
 
         $this->ajaxGetMetadata($origin, $node);
 
-        Toast::send($this->__('communityheader.subscribed'));
+        $this->toast($this->__('communityheader.subscribed'));
     }
 
     public function onSubscriptionUnsupported(Packet $packet)
     {
-        Toast::send($this->__('communityheader.subscription_unsupported'));
+        $this->toast($this->__('communityheader.subscription_unsupported'));
     }
 
     public function onUnsubscribed(Packet $packet)
@@ -75,7 +76,7 @@ class CommunityHeader extends Base
 
         $this->ajaxGetMetadata($origin, $node);
 
-        Toast::send($this->__('communityheader.unsubscribed'));
+        $this->toast($this->__('communityheader.unsubscribed'));
     }
 
     public function ajaxGetMetadata($origin, $node)