+++ app/Widgets/Onboarding/Onboarding.php
@@ -4,7 +4,6 @@ namespace App\Widgets\Onboarding;
 
 use Movim\Widget\Base;
 use App\Widgets\Dialog\Dialog;
-use App\Widgets\Toast\Toast;
 
 class Onboarding extends Base
 {
@@ -19,7 +18,7 @@ class Onboarding extends Base
         $tpl = $this->tpl();
         $this->rpc('Onboarding.setPublic');
 
-        if ($this->me->public == null) {
+        if ($this->me->public === null) {
             Dialog::fill($tpl->draw('_onboarding_public'));
         }
     }
@@ -27,12 +26,12 @@ class Onboarding extends Base
     public function ajaxEnablePublic()
     {
         $this->me->setPublic();
-        Toast::send($this->__('vcard.public'));
+        $this->toast($this->__('profile.public'));
     }
 
     public function ajaxEnableRestricted()
     {
         $this->me->setPrivate();
-        Toast::send($this->__('vcard.restricted'));
+        $this->toast($this->__('profile.restricted'));
     }
 }