+++ src/Moxl/Xec/Action/Avatar/Set.php
@@ -22,10 +22,17 @@ class Set extends Action
 
         if ($this->_url === false) {
             Avatar::set($this->_data, $this->_to, $this->_node, $this->_withPublishOption);
+        } else {
+            // For an URL we simply set the Metadata
+            $setMetadata = new SetMetadata;
+            $setMetadata->setTo($this->_to)
+                ->setNode($this->_node)
+                ->setUrl($this->_url)
+                ->setData($this->_data)
+                ->setWidthMetadata($this->_widthMetadata)
+                ->setHeightMetadata($this->_heightMetadata)
+                ->request();
         }
-
-        Avatar::setMetadata($this->_data, $this->_url, $this->_to, $this->_node,
-            $this->_widthMetadata, $this->_heightMetadata, $this->_withPublishOption);
     }
 
     public function setWidthMetadata($width)
@@ -42,18 +49,31 @@ class Set extends Action
 
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
+        $setMetadata = new SetMetadata;
+        $setMetadata->setTo($this->_to)
+            ->setNode($this->_node)
+            ->setUrl($this->_url)
+            ->setData($this->_data)
+            ->setWidthMetadata($this->_widthMetadata)
+            ->setHeightMetadata($this->_heightMetadata)
+            ->request();
+
         if ($this->_to == false && $this->_node == false) {
             $me = me()->contact;
-            $me->avatartype = 'urn:xmpp:avatar:metadata';
+            $me->avatartype = Avatar::NODE_METADATA;
             $me->save();
 
-            $this->pack($me);
             $this->deliver();
         } else {
             $this->method('pubsub');
             $this->pack(['to' => $this->_to, 'node' => $this->_node]);
             $this->deliver();
         }
+    }
+
+    public function errorPayloadTooBig(string $errorId, ?string $message = null)
+    {
+        $this->deliver();
     }
 
     public function errorFeatureNotImplemented(string $errorId, ?string $message = null)
@@ -84,9 +104,9 @@ class Set extends Action
     public function errorConflict(string $errorId, ?string $message = null)
     {
         $config = new SetConfig;
-        $config->setNode(Avatar::$nodeData)
-               ->setData(Avatar::$nodeConfig)
-               ->request();
+        $config->setNode(Avatar::NODE_DATA)
+            ->setData(Avatar::$nodeConfig)
+            ->request();
 
         $this->_withPublishOption = false;
         $this->request();